LPD-100558 Move the class name savepoint utility to portal-test
DataCleanupTestUtil turned the ClassName_ snapshot-and-delete idiom into a SafeCloseable savepoint so the seven data cleanup preupgrade tests could spell it once instead of seven times. It landed inside portal-upgrade-test's own testIntegration source set, which no other module can depend on, so every test outside that package still has to hand-roll the same code.

The savepoints exist to satisfy @DataGuard(autoDelete = false) and they reach for kernel services only, so portal-test is where they belong. Moving them to com.liferay.portal.kernel.test.util puts them in the same bundle as DataGuard and DataGuardTestRule and makes them reachable from every integration test without a build.gradle change.
